rename to upgrade_from_current_contract
