+++ crates/contract/tests/sandbox/mod.rs
@@ -2,7 +2,7 @@ pub mod ckd;
 pub mod integration_tee_cleanup_after_resharing;
 pub mod sign;
 pub mod tee;
-pub mod updates;
+pub mod upgrade_from_current_contract;
 pub mod upgrade_to_current_contract;
 pub mod vote;
